test: Test all possible type, description and payload lengths to add_key
Test all possible type, description and payload lengths to add_key() to
make sure that the kernel doesn't crash when handling them.

The bulk of this test is implemented in C in the keyctl command so that it
completes in a reasonable amount of time (testing over a million different
sizes of payload from shell script is just too slow).

Signed-off-by: David Howells <dhowells@redhat.com>