Fix the guess for the default LIBDIR
The guesser for the default LIBDIR examines the libraries used by make,
looking for libc.  Make uses libcrypt as well as libc, so if libcrypt
happens to come first, that matches and the sed command to extract the
library path fails.

Fix this by looking for a match on "libc." instead of "libc".

Signed-off-by: David Howells <dhowells@redhat.com>